# Conflicts:
#	Source/EventFlow.Hangfire.Tests/EventFlow.Hangfire.Tests.csproj
#	Source/EventFlow.Hangfire.Tests/paket.references
#	Source/EventFlow.MsSql.Tests/EventFlow.MsSql.Tests.csproj
#	Source/EventFlow.MsSql.Tests/paket.references
#	Source/EventFlow.TestHelpers/EventFlow.TestHelpers.csproj
#	paket.dependencies
#	paket.lock

+++ Source/EventFlow.EventStores.EventStore/EventStoreEventPersistence.cs
@@ -172,7 +172,7 @@ namespace EventFlow.EventStores.EventStore
                     200,
                     false)
                     .ConfigureAwait(false);
-                nextSliceStart = currentSlice.NextEventNumber;
+                nextSliceStart = (int)currentSlice.NextEventNumber;
                 streamEvents.AddRange(currentSlice.Events);
 
             }
@@ -191,7 +191,7 @@ namespace EventFlow.EventStores.EventStore
             return resolvedEvents
                 .Select(e => new EventStoreEvent
                     {
-                        AggregateSequenceNumber = e.Event.EventNumber + 1, // Starts from zero
+                        AggregateSequenceNumber = (int)(e.Event.EventNumber + 1), // Starts from zero
                         Metadata = Encoding.UTF8.GetString(e.Event.Metadata),
                         AggregateId = e.OriginalStreamId,
                         Data = Encoding.UTF8.GetString(e.Event.Data),
